improve the services edit at provider side

+++ app/src/main/java/com/example/mlej/EditServicesActivity.java
@@ -17,7 +17,9 @@ public class EditServicesActivity extends AppCompatActivity {
     SharedPreferences preferences;
     int userId;
 
+    boolean servicesModified = false;
 
+    Button btnUpdateServices;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -38,7 +40,7 @@ public class EditServicesActivity extends AppCompatActivity {
         CheckBox cboxServices8 = findViewById(R.id.cboxService8);
         CheckBox cboxServices9 = findViewById(R.id.cboxService9);
         CheckBox cboxServices10 = findViewById(R.id.cboxService10);
-        Button btnUpdateServices = findViewById(R.id.btnUpdateServices);
+        btnUpdateServices = findViewById(R.id.btnUpdateServices);
 
         if (db.hasServices(userId, 1)) cboxServices1.setChecked(true);
         if (db.hasServices(userId, 2)) cboxServices2.setChecked(true);
@@ -54,21 +56,31 @@ public class EditServicesActivity extends AppCompatActivity {
         btnUpdateServices.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                for( int i=1; i<11; i++) db.removeServices(userId, i);
-                if(cboxServices1.isChecked()) db.addProviderServices(userId, 1);
-                if(cboxServices2.isChecked()) db.addProviderServices(userId, 2);
-                if(cboxServices3.isChecked()) db.addProviderServices(userId, 3);
-                if(cboxServices4.isChecked()) db.addProviderServices(userId, 4);
-                if(cboxServices5.isChecked()) db.addProviderServices(userId, 5);
-                if(cboxServices6.isChecked()) db.addProviderServices(userId, 6);
-                if(cboxServices7.isChecked()) db.addProviderServices(userId, 7);
-                if(cboxServices8.isChecked()) db.addProviderServices(userId, 8);
-                if(cboxServices9.isChecked()) db.addProviderServices(userId, 9);
-                if(cboxServices10.isChecked()) db.addProviderServices(userId, 10);
+                if(!servicesModified) finish();
+                else{
+                    for( int i=1; i<11; i++) db.removeServices(userId, i);
+                    if(cboxServices1.isChecked()) db.addProviderServices(userId, 1);
+                    if(cboxServices2.isChecked()) db.addProviderServices(userId, 2);
+                    if(cboxServices3.isChecked()) db.addProviderServices(userId, 3);
+                    if(cboxServices4.isChecked()) db.addProviderServices(userId, 4);
+                    if(cboxServices5.isChecked()) db.addProviderServices(userId, 5);
+                    if(cboxServices6.isChecked()) db.addProviderServices(userId, 6);
+                    if(cboxServices7.isChecked()) db.addProviderServices(userId, 7);
+                    if(cboxServices8.isChecked()) db.addProviderServices(userId, 8);
+                    if(cboxServices9.isChecked()) db.addProviderServices(userId, 9);
+                    if(cboxServices10.isChecked()) db.addProviderServices(userId, 10);
 
-                Toast.makeText(EditServicesActivity.this, "Services updated successfully", Toast.LENGTH_SHORT).show();
+                    Toast.makeText(EditServicesActivity.this, "Services updated successfully", Toast.LENGTH_SHORT).show();
+                    btnUpdateServices.setText("Close");
+                    servicesModified = false;
+                }
             }
         });
 
+    }
+
+    public void setServicesModified(View v){
+        servicesModified = true;
+        btnUpdateServices.setText("Update the services");
     }
 }
